+++ crates/stackable-operator/src/eos/mod.rs
@@ -27,6 +27,10 @@ pub struct EndOfSupportOptions {
     ))]
     pub interval: Duration,
 
+    /// If the end-of-support check should be disabled entirely.
+    #[cfg_attr(feature = "clap", arg(long = "eos-disabled", env = "EOS_DISABLED"))]
+    pub disabled: bool,
+
     /// The support duration (how long the operator should be considered supported after
     /// it's built-date).
     ///
@@ -65,6 +69,7 @@ pub enum Error {
 pub struct EndOfSupportChecker {
     datetime: DateTime<Utc>,
     interval: Duration,
+    disabled: bool,
 }
 
 impl EndOfSupportChecker {
@@ -79,6 +84,7 @@ impl EndOfSupportChecker {
         let EndOfSupportOptions {
             interval,
             support_duration,
+            disabled,
             ..
         } = options;
 
@@ -95,7 +101,11 @@ impl EndOfSupportChecker {
         // Add the support duration to the built date. This marks the end-of-support date.
         datetime += *support_duration;
 
-        Ok(Self { datetime, interval })
+        Ok(Self {
+            datetime,
+            interval,
+            disabled,
+        })
     }
 
     /// Run the end-of-support checker.
@@ -103,6 +113,11 @@ impl EndOfSupportChecker {
     /// It is recommended to run the end-of-support checker via [`futures::try_join!`] or
     /// [`tokio::join`] alongside other futures (eg. for controllers).
     pub async fn run(self) {
+        // Immediately return if the end-of-support checker is disabled.
+        if self.disabled {
+            return;
+        }
+
         // Construct an interval which can be polled.
         let mut interval = tokio::time::interval(self.interval.into());